project feature: add express-validator

+++ server/project/contacts/contacts.controller.js
@@ -10,12 +10,18 @@ module.exports = function  () {
   }
   
   function getContact (req, res) {
+    req.checkParams('id').isMongoId();
+    if (req.validationErrors()) return res.status(400).send('Bad request');
+    
     Contact.findOne({_id: req.params.id}).exec()
       .then(contact => res.status(200).send(contact))
       .catch(err => res.status(500).send('Unable to access contact'));
   }
   
   function remove (req, res) {
+    req.checkParams('id').isMongoId();
+    if (req.validationErrors()) return res.status(400).send('Bad request');
+    
     Contact.findByIdAndRemove(req.params.id).exec()
       .then(data => res.status(200).send('Contact successfully removed'))
       .catch(err => res.status(500).send('Unable to remove contact'));
@@ -29,6 +35,20 @@ module.exports = function  () {
   }
   
   function save (req, res) {
+    req.sanitizeBody('name').toString();
+    req.sanitizeBody('surname').toString();
+    req.sanitizeBody('phone').toString();
+    req.sanitizeBody('email').toString();
+    req.sanitizeBody('img').toString();
+    
+    req.checkBody('name').isLength({min: 1, max: undefined});
+    req.checkBody('surname').isLength({min: 1, max: undefined});
+    req.checkBody('phone').isNumeric().isLength({min: 9, max: 10});
+    if (req.body.email) req.checkBody('email').isEmail();
+    if (req.body.img) req.checkBody('img').isURL();
+    
+    if (req.validationErrors()) return res.status(400).send('Nice try David!');
+    
     var defImg = 'http://www.golenbock.com/wp-content/uploads/2015/01/placeholder-user-400x400.png';
     req.body.img = req.body.img || defImg;
     var newContact = new Contact(req.body);
@@ -39,6 +59,23 @@ module.exports = function  () {
   }
   
   function update (req, res) {
+    req.checkParams('id').isMongoId();
+    if (req.validationErrors()) return res.status(400).send('Bad request');
+    
+    req.sanitizeBody('name').toString();
+    req.sanitizeBody('surname').toString();
+    req.sanitizeBody('phone').toString();
+    req.sanitizeBody('email').toString();
+    req.sanitizeBody('img').toString();
+    
+    req.checkBody('name').isLength({min: 1, max: undefined});
+    req.checkBody('surname').isLength({min: 1, max: undefined});
+    req.checkBody('phone').isNumeric().isLength({min: 9, max: 10});
+    if (req.body.email) req.checkBody('email').isEmail();
+    if (req.body.img) req.checkBody('img').isURL();
+    
+    if (req.validationErrors()) return res.status(400).send('Nice try David!');
+    
     Contact.findByIdAndUpdate(req.params.id, {$set: req.body})
       .then(data => res.status(200).send('Contact successfully updated'))
       .catch(err => res.status(500).send('Unable to update contact'))
